==> research/simulation/script/0b_sim_functions.R <==
..get_jobsII_formulas <- function(outcome) {

    X.names <- c("age", "sex", "race", "edu", "marital", "hh.kids",
                 "hh.income", "econ.hard", "occu", "wks.unemp",
                 "part.motiv", "seek.motiv", "seek.effi",
                 "assertive", "depress1")

    X.cont <- c("age", "econ.hard", "wks.unemp", "part.motiv", "seek.motiv",
                "seek.effi", "assertive", "depress1")
    X.square <- paste0(paste0("I("     , X.cont), "^2)")
    X.sqroot <- paste0(paste0("I(sqrt(", X.cont), "))" )
    X.all <- c(X.names, X.square, X.sqroot)

    Z.form <- paste("z ~ ", paste(X.all,   collapse = " + "))
    C.form <- paste("c ~ ", paste(X.all,   collapse = " + "))
    Y.form <- paste("y ~ ", paste(X.names, collapse = " + "))


    mget(c("Z.form", "C.form", "Y.form", "X.names"))


}




..get_true_params <- function(seed = 123) {

    master <- list(num = 20,
                   size = 1000000,
                   seed = seed)

    set.seed(master$seed)
    seeds <- sample(1:.Machine$integer.max, master$num)

    mget(c("master", "seeds"))
}





..get_sim_params <- function(seed = 987) {

    master <- list(num = 500,
                   size = 465, # size of analytic data
                   seed = seed)

    set.seed(master$seed)
    seeds <- sample(1:.Machine$integer.max, master$num)

    mget(c("master", "seeds"))
}





..process_sim_results <- function(sim.results, true.results) {

    true.values <- true.results$values

    results <- sim.results$results

    results$bc1 <- results$bm1 + 2 * (results$raw - results$bm1)
    results$bc2 <- results$bm2 + 3 * (results$raw - results$bm1)

    results$error0 <- results$raw - replicate(dim(results$raw)[3], true.results$values, simplify = "array")
    results$error1 <- results$bc1 - replicate(dim(results$bc1)[3], true.results$values, simplify = "array")
    results$error2 <- results$bc2 - replicate(dim(results$bc2)[3], true.results$values, simplify = "array")

    sim.results$results <- results

    sim.results

}






..get_sim_basis_work <- function() {

    X.names <- c("age", "sex", "race", "edu", "marital", "hh.kids",
                 "hh.income", "econ.hard", "occu", "wks.unemp",
                 "part.motiv", "seek.motiv", "seek.effi",
                 "assertive", "depress1")
    X.cont <- c("age", "econ.hard", "wks.unemp", "part.motiv", "seek.motiv",
                "seek.effi", "assertive", "depress1")
    X.square <- paste0(paste0("I("     , X.cont), "^2)")
    X.sqroot <- paste0(paste0("I(sqrt(", X.cont), "))" )
    X.all    <- c(X.names, X.square, X.sqroot)


    dat <- readRDS(here::here("data-raw", "dat.rds"))

    # base data for simulating (X,Z)
    xz.base       <- dat[c(X.names, "treat")]
    xz.base$treat <- factor(xz.base$treat)


    # model for simulating C
    c.form <- paste("complier ~", paste(X.all, collapse = " + "))
    c.mod  <- glm(formula = c.form, family = "binomial", data = dat[dat$treat==1,])

    # models for simulating Y part 1 (work3)
    y.form <- paste("work3 ~", paste(X.names, collapse = " + "))
    y0.mod  <- glm(formula = y.form, family = "binomial", data = dat[dat$treat==0,])
    y1c.mod <- glm(formula = y.form, family = "binomial", data = dat[dat$treat==1 & dat$complier==1,])
    y1n.mod <- glm(formula = y.form, family = "binomial", data = dat[dat$treat==1 & dat$complier==0,])

    mget(c("xz.base",
           "c.mod",
           "y0.mod", "y1c.mod", "y1n.mod"))
}





..sim_work <- function(xz.base,
                       c.mod,
                       y0.mod,
                       y1c.mod,
                       y1n.mod,
                       size,
                       seed) {

    sim <- synthpop::syn(xz.base, seed = seed, k = size)$syn
    sim$treat <- as.numeric(sim$treat) - 1

    set.seed(seed)

    c.sim <- predict(c.mod, newdata = sim, type = "response")
    c.sim <- rbinom(n = nrow(sim), size = 1, prob = c.sim)

    y0.sim <- predict(y0.mod, newdata = sim, type = "response")
    y0.sim <- rbinom(n = nrow(sim), size = 1, prob = y0.sim)

    y1c.sim <- predict(y1c.mod, newdata = sim, type = "response")
    y1c.sim <- rbinom(n = nrow(sim), size = 1, prob = y1c.sim)

    y1n.sim <- predict(y1n.mod, newdata = sim, type = "response")
    y1n.sim <- rbinom(n = nrow(sim), size = 1, prob = y1n.sim)

    sim$complier <- ifelse(sim$treat==0, NA, c.sim)
    sim$work3    <- ifelse(sim$treat==0, y0.sim,
                           ifelse(sim$complier==1, y1c.sim, y1n.sim))

    sim
}




..get_sim_basis_earn <- function() {

    X.names <- c("age", "sex", "race", "edu", "marital", "hh.kids",
                 "hh.income", "econ.hard", "occu", "wks.unemp",
                 "part.motiv", "seek.motiv", "seek.effi",
                 "assertive", "depress1")
    X.cont <- c("age", "econ.hard", "wks.unemp", "part.motiv", "seek.motiv",
                "seek.effi", "assertive", "depress1")
    X.square <- paste0(paste0("I("     , X.cont), "^2)")
    X.sqroot <- paste0(paste0("I(sqrt(", X.cont), "))" )
    X.all    <- c(X.names, X.square, X.sqroot)


    dat <- readRDS(here::here("data-raw", "dat.rds"))

    # base data for simulating (X,Z)
    xz.base       <- dat[c(X.names, "treat")]
    xz.base$treat <- factor(xz.base$treat)


    # model for simulating C
    c.form <- paste("complier ~", paste(X.all, collapse = " + "))
    c.mod  <- glm(formula = c.form, family = "binomial", data = dat[dat$treat==1,])

    # models for simulating Y part 1 (work3)
    y.form.w <- paste("work3 ~", paste(X.names, collapse = " + "))
    y0.mod.w  <- glm(formula = y.form.w, family = "binomial", data = dat[dat$treat==0,])
    y1c.mod.w <- glm(formula = y.form.w, family = "binomial", data = dat[dat$treat==1 & dat$complier==1,])
    y1n.mod.w <- glm(formula = y.form.w, family = "binomial", data = dat[dat$treat==1 & dat$complier==0,])

    # models for simulating Y part 2 (earn3)
    y.form.e   <- paste("earn3 ~", paste(X.names, collapse = " + "))
    y.bounds.e <- c(0, 6000)
    dat$earn3  <- .trans_bounds(dat$earn3, from = y.bounds.e, to = 0:1)

    y0.mod.e  <- glm(formula = y.form.e, family = "quasibinomial", data = dat[dat$treat==0 & dat$work3==1,])
    y1c.mod.e <- glm(formula = y.form.e, family = "quasibinomial", data = dat[dat$treat==1 & dat$work3==1 & dat$complier==1,])
    y1n.mod.e <- glm(formula = y.form.e, family = "quasibinomial", data = dat[dat$treat==1 & dat$work3==1 & dat$complier==0,])

    y0.dispersion  <- .extract_dispersion_wtd_glm(y0.mod.e)
    y1c.dispersion <- .extract_dispersion_wtd_glm(y1c.mod.e)
    y1n.dispersion <- .extract_dispersion_wtd_glm(y1n.mod.e)

    y0.psi.e  <- (1 - y0.dispersion)  / y0.dispersion   # computing precision
    y1c.psi.e <- (1 - y1c.dispersion) / y1c.dispersion
    y1n.psi.e <- (1 - y1n.dispersion) / y1n.dispersion

    y.bounds.e

    mget(c("xz.base",
           "c.mod",
           "y0.mod.w", "y1c.mod.w", "y1n.mod.w",
           "y0.mod.e", "y1c.mod.e", "y1n.mod.e",
           "y0.psi.e", "y1c.psi.e", "y1n.psi.e",
           "y.bounds.e"))
}




..sim_earn <- function(xz.base,
                       c.mod,
                       y0.mod.w,
                       y1c.mod.w,
                       y1n.mod.w,
                       y0.mod.e,
                       y1c.mod.e,
                       y1n.mod.e,
                       y0.psi.e,
                       y1c.psi.e,
                       y1n.psi.e,
                       y.bounds.e,
                       size,
                       seed) {

    sim <- synthpop::syn(xz.base, seed = seed, k = size)$syn
    sim$treat <- as.numeric(sim$treat) - 1

    set.seed(seed)

    # simulate complier
    c.sim <- predict(c.mod, newdata = sim, type = "response")
    c.sim <- rbinom(n = nrow(sim), size = 1, prob = c.sim)

    sim$complier <- ifelse(sim$treat==0, NA, c.sim)

    # simulate work3 before earn3
    y0.sim  <- predict(y0.mod.w,  newdata = sim, type = "response")
    y1c.sim <- predict(y1c.mod.w, newdata = sim, type = "response")
    y1n.sim <- predict(y1n.mod.w, newdata = sim, type = "response")

    y0.sim  <- rbinom(n = nrow(sim), size = 1, prob = y0.sim)
    y1c.sim <- rbinom(n = nrow(sim), size = 1, prob = y1c.sim)
    y1n.sim <- rbinom(n = nrow(sim), size = 1, prob = y1n.sim)

    sim$work3 <- ifelse(sim$treat==0, y0.sim,
                        ifelse(sim$complier==1, y1c.sim, y1n.sim))

    rm(y0.sim, y1c.sim, y1n.sim)

    # simulate earn3
    y0.sim  <- predict(y0.mod.e,  newdata = sim, type = "response")
    y1c.sim <- predict(y1c.mod.e, newdata = sim, type = "response")
    y1n.sim <- predict(y1n.mod.e, newdata = sim, type = "response")

    y0.sim  <- rbeta(n = nrow(sim), shape1 = y0.sim *y0.psi.e,  shape2 = (1-y0.sim) *y0.psi.e)
    y1c.sim <- rbeta(n = nrow(sim), shape1 = y1c.sim*y1c.psi.e, shape2 = (1-y1c.sim)*y1c.psi.e)
    y1n.sim <- rbeta(n = nrow(sim), shape1 = y1n.sim*y1n.psi.e, shape2 = (1-y1n.sim)*y1n.psi.e)

    sim$earn3 <- ifelse(sim$work3==0, 0,
                        ifelse(sim$treat==0, y0.sim,
                               ifelse(sim$complier==1, y1c.sim, y1n.sim)))

    sim$earn3 <- .trans_bounds(sim$earn3, from = 0:1, to = y.bounds.e)

    sim
}










..get_sim_basis_depress <- function() {

    X.names <- c("age", "sex", "race", "edu", "marital", "hh.kids",
                 "hh.income", "econ.hard", "occu", "wks.unemp",
                 "part.motiv", "seek.motiv", "seek.effi",
                 "assertive", "depress1")
    X.cont <- c("age", "econ.hard", "wks.unemp", "part.motiv", "seek.motiv",
                "seek.effi", "assertive", "depress1")
    X.square <- paste0(paste0("I("     , X.cont), "^2)")
    X.sqroot <- paste0(paste0("I(sqrt(", X.cont), "))" )
    X.all    <- c(X.names, X.square, X.sqroot)


    dat <- readRDS(here::here("data-raw", "dat.rds"))

    # base data for simulating (X,Z)
    xz.base       <- dat[c(X.names, "treat")]
    xz.base$treat <- factor(xz.base$treat)


    # model for simulating C
    c.form <- paste("complier ~", paste(X.all, collapse = " + "))
    c.mod  <- glm(formula = c.form, family = "binomial", data = dat[dat$treat==1,])

    # models for simulating Y part 1 (work3)
    y.form <- paste("depress3 ~", paste(X.names, collapse = " + "))
    y.bounds <- c(1, 5)
    dat$depress3 <- .trans_bounds(dat$depress3, from = y.bounds, to = 0:1)

    y0.mod  <- glm(formula = y.form, family = "quasibinomial", data = dat[dat$treat==0,])
    y1c.mod <- glm(formula = y.form, family = "quasibinomial", data = dat[dat$treat==1 & dat$complier==1,])
    y1n.mod <- glm(formula = y.form, family = "quasibinomial", data = dat[dat$treat==1 & dat$complier==0,])

    y0.dispersion  <- .extract_dispersion_wtd_glm(y0.mod)
    y1c.dispersion <- .extract_dispersion_wtd_glm(y1c.mod)
    y1n.dispersion <- .extract_dispersion_wtd_glm(y1n.mod)

    y0.psi  <- (1 - y0.dispersion)  / y0.dispersion   # computing precision
    y1c.psi <- (1 - y1c.dispersion) / y1c.dispersion
    y1n.psi <- (1 - y1n.dispersion) / y1n.dispersion

    mget(c("xz.base",
           "c.mod",
           "y0.mod", "y1c.mod", "y1n.mod",
           "y0.psi", "y1c.psi", "y1n.psi",
           "y.bounds"))
}




..sim_depress <- function(xz.base,
                          c.mod,
                          y0.mod,
                          y1c.mod,
                          y1n.mod,
                          y0.psi,
                          y1c.psi,
                          y1n.psi,
                          y.bounds,
                          size,
                          seed) {

    sim <- synthpop::syn(xz.base, seed = seed, k = size)$syn
    sim$treat <- as.numeric(sim$treat) - 1

    set.seed(seed)

    c.sim <- predict(c.mod, newdata = sim, type = "response")
    c.sim <- rbinom(n = nrow(sim), size = 1, prob = c.sim)

    sim$complier <- ifelse(sim$treat==0, NA, c.sim)

    y0.sim  <- predict(y0.mod,  newdata = sim, type = "response")
    y1c.sim <- predict(y1c.mod, newdata = sim, type = "response")
    y1n.sim <- predict(y1n.mod, newdata = sim, type = "response")

    y0.sim  <- rbeta(n = nrow(sim), shape1 = y0.sim *y0.psi,  shape2 = (1-y0.sim) *y0.psi)
    y1c.sim <- rbeta(n = nrow(sim), shape1 = y1c.sim*y1c.psi, shape2 = (1-y1c.sim)*y1c.psi)
    y1n.sim <- rbeta(n = nrow(sim), shape1 = y1n.sim*y1n.psi, shape2 = (1-y1n.sim)*y1n.psi)

    sim$depress3 <- ifelse(sim$treat==0, y0.sim,
                           ifelse(sim$complier==1, y1c.sim, y1n.sim))

    sim$depress3 <- .trans_bounds(sim$depress3, from = 0:1, to = y.bounds)

    sim
}




..plot_sim_bias <- function(sim,
                            sens.type,
                            ylimits,
                            ylimits.std,
                            include.Y1 = TRUE) {


    error0 <- sim$results$error0
    error1 <- sim$results$error1
    error2 <- sim$results$error2
    sens.params <- sim$sens.params
    rm(sim)

    # true.se <- true$se
    # rm(true)


    if (sens.type %in% c("OR", "GOR", "MR")) {
        x.ticks <- unique(round(sens.params[,"sens.para.coord"]))
        x.tick.labels <- ifelse(x.ticks==0, as.character(1),
                                ifelse(x.ticks>0, as.character(x.ticks+1),
                                       paste0("1/", as.character(abs(x.ticks)+1))))

        x.label <- paste("sensitivity", sens.type)
    }

    if (sens.type %in% c("SMDe", "SMD")) {
        x.ticks <- unique(round(sens.params[,"sens.para.coord"] * 2)/2)
        x.tick.labels <- x.ticks

        x.label <- "sensitivity SMD"
    }

    color.values <- c("black", "blue", "dodgerblue")

    all.estimands <- c("cace", "nace", "tau.0c", "tau.0n", "tau.1c", "tau.1n")
    c.estimands   <- c("cace", "tau.0c", "tau.1c")
    effects       <- c("cace", "nace")




    ########################################
    # BIAS PLOT

    bias0 <- apply(error0, c(1,2), function(z) mean(z, na.rm = TRUE))
    bias1 <- apply(error1, c(1,2), function(z) mean(z, na.rm = TRUE))
    bias2 <- apply(error2, c(1,2), function(z) mean(z, na.rm = TRUE))

    dat <- rbind(cbind(data.frame(cbind(sens.params, bias0)), estimate = "uncorrected"),
                 cbind(data.frame(cbind(sens.params, bias1)), estimate = "BC1"),
                 cbind(data.frame(cbind(sens.params, bias2)), estimate = "BC2"))
    dat$estimate <- factor(dat$estimate, levels = c("uncorrected", "BC1", "BC2"))

    dat <- tidyr::pivot_longer(dat,
                               cols = all_of(all.estimands),
                               names_to = "estimand",
                               values_to = "metric")
    dat$stratum <- ifelse(dat$estimand %in% c.estimands, "complier", "noncomplier")
    dat$stratum <- factor(dat$stratum, levels = c("noncomplier", "complier"))
    dat$estimand <- ifelse(dat$estimand %in% effects, "average effect",
                           ifelse(dat$estimand %in% c("tau.0c", "tau.0n"), "mean Y0", "mean Y1"))

    if (!include.Y1) dat <- dat[dat$estimand!="mean Y1",]




    # se <- data.frame(cbind(sens.params, true.se))
    # se <- tidyr::pivot_longer(se,
    #                           cols = all.estimands,
    #                           names_to = "estimand",
    #                           values_to = "metric")
    # se$stratum <- ifelse(se$estimand %in% c.estimands, "complier", "noncomplier")
    # se$stratum <- factor(se$stratum, levels = c("noncomplier", "complier"))
    # se$estimand <- ifelse(se$estimand %in% effects, "average effect",
    #                       ifelse(se$estimand %in% c("tau.0c", "tau.0n"), "mean Y0", "mean Y1"))
    #
    # if (!include.Y1) se <- se[se$estimand!="mean Y1",]
    #
    # se.minus <- se
    # se.minus$metric <- -se.minus$metric
    # se <- rbind(cbind(se,       pos = "above"),
    #             cbind(se.minus, pos = "below"))
    # rm(se.minus)




    p.bias <-
        ggplot2::ggplot(dat,
                        ggplot2::aes(x = sens.para.coord,
                                     y = metric,
                                     color = estimate)) +
        ggplot2::geom_hline(yintercept = 0, color = "lightgray") +
        ggplot2::geom_vline(xintercept = 0, color = "lightgray") +
        # ggplot2::geom_line(data = se, ggplot2::aes(group = pos), color = "gray70") +
        ggplot2::geom_line() +
        ggplot2::scale_x_continuous(breaks = x.ticks, labels = x.tick.labels) +
        ggplot2::scale_color_manual(name = "", values = color.values) +
        ggplot2::labs(x = x.label, y = "", title = "Bias") +
        ggplot2::facet_grid(estimand ~ stratum) +
        ggplot2::theme_bw()

    if (!missing(ylimits))
        p.bias <- p.bias +
        ggplot2::coord_cartesian(ylim = ylimits)






    ########################################
    # STANDARDIZED BIAS PLOT

    se0 <- apply(error0, c(1,2), function(z) sd(z, na.rm = TRUE))
    se1 <- apply(error1, c(1,2), function(z) sd(z, na.rm = TRUE))
    se2 <- apply(error2, c(1,2), function(z) sd(z, na.rm = TRUE))

    bias0 <- bias0 / se0
    bias1 <- bias1 / se0
    bias2 <- bias2 / se0



    dat <- rbind(cbind(data.frame(cbind(sens.params, bias0)), estimate = "uncorrected"),
                 cbind(data.frame(cbind(sens.params, bias1)), estimate = "BC1"),
                 cbind(data.frame(cbind(sens.params, bias2)), estimate = "BC2"))
    dat$estimate <- factor(dat$estimate, levels = c("uncorrected", "BC1", "BC2"))

    dat <- tidyr::pivot_longer(dat,
                               cols = all_of(all.estimands),
                               names_to = "estimand",
                               values_to = "metric")
    dat$stratum <- ifelse(dat$estimand %in% c.estimands, "complier", "noncomplier")
    dat$stratum <- factor(dat$stratum, levels = c("noncomplier", "complier"))
    dat$estimand <- ifelse(dat$estimand %in% effects, "average effect",
                           ifelse(dat$estimand %in% c("tau.0c", "tau.0n"), "mean Y0", "mean Y1"))

    if (!include.Y1) dat <- dat[dat$estimand!="mean Y1",]




    # se <- data.frame(cbind(sens.params, true.se / se0))
    # se <- tidyr::pivot_longer(se,
    #                           cols = all.estimands,
    #                           names_to = "estimand",
    #                           values_to = "metric")
    # se$stratum <- ifelse(se$estimand %in% c.estimands, "complier", "noncomplier")
    # se$stratum <- factor(se$stratum, levels = c("noncomplier", "complier"))
    # se$estimand <- ifelse(se$estimand %in% effects, "average effect",
    #                       ifelse(se$estimand %in% c("tau.0c", "tau.0n"), "mean Y0", "mean Y1"))
    #
    # if (!include.Y1) se <- se[se$estimand!="mean Y1",]
    #
    # se.minus <- se
    # se.minus$metric <- -se.minus$metric
    # se <- rbind(cbind(se,       pos = "above"),
    #             cbind(se.minus, pos = "below"))
    # rm(se.minus)


    p.std.bias <-
        ggplot2::ggplot(dat,
                        ggplot2::aes(x = sens.para.coord,
                                     y = metric,
                                     color = estimate)) +
        ggplot2::geom_hline(yintercept = 0, color = "lightgray") +
        ggplot2::geom_vline(xintercept = 0, color = "lightgray") +
        # ggplot2::geom_line(data = se, ggplot2::aes(group = pos), color = "gray70") +
        ggplot2::geom_line() +
        ggplot2::scale_x_continuous(breaks = x.ticks, labels = x.tick.labels) +
        ggplot2::scale_color_manual(name = "", values = color.values) +
        ggplot2::labs(x = x.label, y = "", title = "Standardized bias") +
        ggplot2::facet_grid(estimand ~ stratum) +
        ggplot2::theme_bw()

    if (!missing(ylimits.std))
        p.std.bias <- p.std.bias  +
        ggplot2::coord_cartesian(ylim = ylimits.std)

    rm(bias0, bias1, bias2)




    ########################################
    # STANDARDIZED SE DIFF PLOT

    se.diff1 <- (se1 - se0) / se0
    se.diff2 <- (se2 - se0) / se0

    rm(se0, se1, se2)

    dat <- rbind(cbind(data.frame(cbind(sens.params, se.diff1)), estimate = "BC1"),
                 cbind(data.frame(cbind(sens.params, se.diff2)), estimate = "BC2"))

    dat$estimate <- factor(dat$estimate, levels = c("BC1", "BC2"))

    dat <- tidyr::pivot_longer(dat,
                               cols = all_of(all.estimands),
                               names_to = "estimand",
                               values_to = "metric")
    dat$stratum <- ifelse(dat$estimand %in% c.estimands,
                          "complier", "noncomplier")
    dat$stratum <- factor(dat$stratum, levels = c("noncomplier", "complier"))
    dat$estimand <- ifelse(dat$estimand %in% c("cace", "nace"),
                           "average effect",
                           ifelse(dat$estimand %in% c("tau.0c", "tau.0n"),
                                  "mean Y0", "mean Y1"))

    if (!include.Y1) dat <- dat[dat$estimand!="mean Y1",]


    p.std.SE.diff <-
        ggplot2::ggplot(dat,
                        ggplot2::aes(x = sens.para.coord,
                                     y = metric,
                                     color = estimate)) +
        ggplot2::geom_hline(yintercept = 0, color = "lightgray") +
        ggplot2::geom_vline(xintercept = 0, color = "lightgray") +
        ggplot2::geom_line() +
        ggplot2::scale_x_continuous(breaks = x.ticks, labels = x.tick.labels) +
        ggplot2::scale_color_manual(name = "", values = color.values[2:3]) +
        ggplot2::labs(x = x.label, y = "", title = "Standardized SE change") +
        ggplot2::facet_grid(estimand ~ stratum) +
        ggplot2::theme_bw()

    if (!missing(ylimits.std))
        p.std.SE.diff <- p.std.SE.diff +
        ggplot2::coord_cartesian(ylim = ylimits.std)



    # Combine into figure
    p.legend <- cowplot::get_legend(
        p.bias +
            ggplot2::guides(color = ggplot2::guide_legend(nrow = 3)))

    cowplot::plot_grid(
        p.bias        + ggplot2::theme(legend.position = "none"),
        p.legend,
        p.std.bias    + ggplot2::theme(legend.position = "none"),
        p.std.SE.diff + ggplot2::theme(legend.position = "none"),
        nrow = 2,
        ncol = 2,
        rel_widths = 1,
        rel_heights = 1
    )
}
